+++ tests/testthat/helper-src.R
@@ -1,0 +1,9 @@
+test_register_src("df", src_df(env = new.env(parent = emptyenv())))
+test_register_src("sqlite", src_sqlite(":memory:", create = TRUE))
+
+if (identical(Sys.info()[["user"]], "hadley")) {
+  test_register_src("postgres", src_postgres("test"))
+} else if (identical(Sys.getenv("TRAVIS"), "true")) {
+  test_register_src("postgres", src_postgres("test", user = "travis", password = ""))
+}
+