+++ tests/testthat/test-mutate.r
@@ -22,8 +22,7 @@ test_that("repeated outputs applied progressively (grouped_df)", {
 })
 
 df <- data.frame(x = 1:10, y = 6:15)
-srcs <- temp_srcs(c("df", "sqlite", "postgres"))
-tbls <- temp_load(srcs, df)
+tbls <- test_load(df)
 
 test_that("two mutates equivalent to one", {
   compare_tbls(tbls, function(tbl) tbl %>% mutate(x2 = x * 2, y4 = y * 4))