+++ tests/testthat/test-select.r
@@ -1,9 +1,7 @@
 context("Select")
 
 df <- as.data.frame(as.list(setNames(1:26, letters)))
-srcs <- temp_srcs(c("df", "sqlite", "postgres"))
-tbls <- temp_load(srcs, df)
-
+tbls <- test_load(df)
 
 test_that("two selects equivalent to one", {
   compare_tbls(tbls, function(tbl) tbl %>% select(l:s) %>% select(n:o),